remove data dir when setup in test mode
