+++ cmd/config.go
@@ -22,14 +22,14 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
-
 	"github.com/ontio/ontology/cmd/utils"
 	"github.com/ontio/ontology/common"
 	"github.com/ontio/ontology/common/config"
+	"github.com/ontio/ontology/common/constants"
 	"github.com/ontio/ontology/common/log"
 	"github.com/ontio/ontology/smartcontract/service/native/governance"
 	"github.com/urfave/cli"
+	"io/ioutil"
 )
 
 func SetOntologyConfig(ctx *cli.Context) (*config.OntologyConfig, error) {
@@ -48,9 +48,9 @@ func SetOntologyConfig(ctx *cli.Context) (*config.OntologyConfig, error) {
 	if cfg.Genesis.ConsensusType == config.CONSENSUS_TYPE_SOLO {
 		cfg.Ws.EnableHttpWs = true
 		cfg.Restful.EnableHttpRestful = true
-		cfg.P2PNode.NetworkId = config.NETWORK_ID_SOLO_NET
-		cfg.P2PNode.NetworkName = config.GetNetworkName(cfg.P2PNode.NetworkId)
-		cfg.P2PNode.NetworkMaigc = config.GetNetworkMagic(cfg.P2PNode.NetworkId)
+		cfg.P2PNode.NetworkId = constants.NETWORK_ID_SOLO_NET
+		cfg.P2PNode.NetworkName = constants.GetNetworkName(cfg.P2PNode.NetworkId)
+		cfg.P2PNode.NetworkMaigc = constants.GetNetworkMagic(cfg.P2PNode.NetworkId)
 	}
 	return cfg, nil
 }
@@ -125,8 +125,8 @@ func setConsensusConfig(ctx *cli.Context, cfg *config.ConsensusConfig) {
 
 func setP2PNodeConfig(ctx *cli.Context, cfg *config.P2PNodeConfig) {
 	cfg.NetworkId = uint32(ctx.GlobalUint(utils.GetFlagName(utils.NetworkIdFlag)))
-	cfg.NetworkMaigc = config.GetNetworkMagic(cfg.NetworkId)
-	cfg.NetworkName = config.GetNetworkName(cfg.NetworkId)
+	cfg.NetworkMaigc = constants.GetNetworkMagic(cfg.NetworkId)
+	cfg.NetworkName = constants.GetNetworkName(cfg.NetworkId)
 	cfg.NodePort = ctx.GlobalUint(utils.GetFlagName(utils.NodePortFlag))
 	cfg.NodeConsensusPort = ctx.GlobalUint(utils.GetFlagName(utils.ConsensusPortFlag))
 	cfg.DualPortSupport = ctx.GlobalBool(utils.GetFlagName(utils.DualPortSupportFlag))
