+++ cmd/utils/flags.go
@@ -20,6 +20,7 @@ package utils
 
 import (
 	"github.com/ontio/ontology/common/config"
+	"github.com/ontio/ontology/common/constants"
 	"github.com/ontio/ontology/smartcontract/service/neovm"
 	"github.com/urfave/cli"
 	"strings"
@@ -104,7 +105,7 @@ var (
 	NetworkIdFlag = cli.UintFlag{
 		Name:  "networkid",
 		Usage: "P2P network id. 1=main net, 2=polaris test net, 3=testmode, and other for custom network",
-		Value: config.NETWORK_ID_POLARIS_NET,
+		Value: constants.NETWORK_ID_POLARIS_NET,
 	}
 	NodePortFlag = cli.UintFlag{
 		Name:  "nodeport",
