+++ common/constants/constants.go
@@ -18,7 +18,10 @@
 
 package constants
 
-import "time"
+import (
+	"fmt"
+	"time"
+)
 
 // genesis constants
 var (
@@ -67,3 +70,40 @@ var UNBOUND_DEADLINE = (func() uint32 {
 
 // multi-sig constants
 const MULTI_SIG_MAX_PUBKEY_SIZE = 1024
+
+const (
+	NETWORK_ID_MAIN_NET      = 1
+	NETWORK_ID_POLARIS_NET   = 2
+	NETWORK_ID_SOLO_NET      = 3
+	NETWORK_NAME_MAIN_NET    = "ontology"
+	NETWORK_NAME_POLARIS_NET = "polaris"
+	NETWORK_NAME_SOLO_NET    = "testmode"
+)
+
+var NETWORK_MAGIC = map[uint32]uint32{
+	NETWORK_ID_MAIN_NET:    0x74746e41, //Network main
+	NETWORK_ID_POLARIS_NET: 0x01015401, //Network polaris
+	NETWORK_ID_SOLO_NET:    0x00000000, //Network solo
+}
+
+var NETWORK_NAME = map[uint32]string{
+	NETWORK_ID_MAIN_NET:    NETWORK_NAME_MAIN_NET,
+	NETWORK_ID_POLARIS_NET: NETWORK_NAME_POLARIS_NET,
+	NETWORK_ID_SOLO_NET:    NETWORK_NAME_SOLO_NET,
+}
+
+func GetNetworkMagic(id uint32) uint32 {
+	nid, ok := NETWORK_MAGIC[id]
+	if ok {
+		return nid
+	}
+	return id
+}
+
+func GetNetworkName(id uint32) string {
+	name, ok := NETWORK_NAME[id]
+	if ok {
+		return name
+	}
+	return fmt.Sprintf("%d", id)
+}
